Created Decide_Percentage_Marks.R
Created a function that given base values, new values, the original plot axis ticks, and a rounding value return a sequence of new ticks. The new ticks describe a second y axis of percent difference values.

For example if the original axis ticks were created from the difference between base values of range 95-100 and new values of range 90-95 a vector of (-10,-8,-6,-4,-2, 0) would be returned.