Updated Adjacent Proximity Criteria
After re-reading zipper (2019) it was clear that the adjacent criteria was not just streams within immediate subwatershed, but also subwatersheds touching the one that the well is within. Added code to reflect this.

In original publication additional adjacent subwatersheds are assigned using a thiessen polygon method, however here to avoid recursion they are assigned using st_touches()