Added conservative estimate to lagged average
Added conservative estimate to lagged average of when a well will start contributing to stream depletions. This avoids including the current timestep in the lagged average, which does not yet contribute.

Weighted mean used to find the starting time to avoid the following scenario:

mean start time is 20, so a lagged average is between t[20:80] when we are at position 100.
A well close to the stream starts to pump a large volume.
The depletions start to affect the stream, but because the mean start time incorporates further wells, the pumping rate in the divisor does not reflect this.
Fractional depletion explodes to a high number despite this not reflecting reality

This is a compromise between complete conservatism (summing t[start:i] where the opposite problem can happen where sudden fluctuations at the current timestep can make the fractional depletion artificially low) and complete liberalism (summing t[start:mean(end)]  where fractional depletion can appear artificially high)