Added geologic apportionment
describes whether to scale apportionment by the ratio between the transmissivity of the geologic unit of the reach and the geologic unit of the well.

The logic behind this is that a reach in more transmissive units will receive more depletion than those in less transmissive units. For example if two reaches are equidistant from a well, and the reaches are in two different geologic units, they should not receive equal apportionment.