Gradient Descent Algorithm to Find SDF Time
Made custom gradient descent algorithm to find the SDF time using the inverse of the analytical models. If using glover model exact inverse is used and gradient descent is neglected. If hunt or hantush used then it finds the time to a user specified multiple of the pumping rate. For example 0.28 will find when the depletions in reach r equal 28% of the pumping rate in well w (this is what the jenkins approximation finds, but using the actual analytical models and not d^2S/T which assumes no clogging layer and infinite aquifer penetration by reach r).